ModuleManager and Code Quality updates
Added migration loading in the module manager; ToDo: fix the issue in which the migrations don't get rolled back when disabling the module.
Renamed improper classes related to the navigation and added a new nested navigation item collection type.
Renamed the improperly named "PanelRoutingServiceProvider" to "NavigationServiceProvider" and renamed the "panelRoutes" shared variable to "cmsSidebarItems" to better reflect the type of content that we're expecting.
Sidebar template in CMS got a tiny rework, in which it loads the navigation item template depending on the type of the class.
Added an informational message to the module list page, in order to let the user know that the modules can't be managed yet through the frontend.