Updates
Added asset bundling in modules.
Added session management from Symfony.
Added model management in the CMS, with working add/edit/new/save functionalities.
Added a core plugin for the CMS which adds styles to be bundled when installing the app.
Views were moved in the core plugin for the CMS.
Tried to fix a styling issue in the base page template.
Reworked the Navigation Item classes and added a new navigation item for models.